Generalize ELF ingestion to ELF32/64, all endianness and architectures
Broaden the container-only ELF parser from its initial ELF32-LE-EM_MIPS
slice to a general bounded reader:

- Parameterize ElfReader by ElfClass{Elf32,Elf64} and ElfEndian{Little,Big}
  with endian-aware u16/u32/u64 reads and a class-sized addr() read; branch
  header/program-header/section-header field layout on class (ELF64 reorders
  p_flags and widens fields).
- Replace the EM_MIPS-only gate with an ElfMachine allow-list (x86, x86-64,
  ARM, AArch64, MIPS/MIPS64, RISC-V, PowerPc/PowerPc64, Other) that still
  parses unknown machines as containers. Accept ET_EXEC and ET_DYN (PIE).
- Recover SHT_SYMTAB/SHT_DYNSYM symbol tables (class-aware 16/24-byte
  records) with string-table name resolution, bounded by MAX_ELF_SYMBOLS
  and a truncation flag; submit STT_FUNC/STT_OBJECT symbols that map into a
  PT_LOAD segment as Function/Global name claims, mirroring the PE pattern.
- Widen ElfProgramHeader/ElfSectionHeader/ElfLoadSegment and ElfAnalysis
  address fields to u64; add class, endian, symbols, symbol_scan_truncated
  with serde defaults so schema<=14 packages still load.
- Derive the architecture string from class+machine+endian (MIPS token
  preserved) and re-validate it deterministically.

Bump package CURRENT_SCHEMA_VERSION 14 -> 15 and extend the CLI decode arm;
new analyses write schema 15 while schemas 1-14 still load.

Also gate the pre-existing ungated RemoteCommandCheckpoint import in
resymbol-windows-debug-host behind cfg(any(windows, test)) so the
full-workspace clippy gate is green.

+++ crates/resymbol-analysis/src/session.rs
@@ -885,8 +885,8 @@ fn validate_elf_subject_mapping(
         .checked_add(size)
         .ok_or(SessionValidationError::ElfSubjectOutsideLoadSegment { index })?;
     let segment = analysis.load_segments.iter().find(|segment| {
-        let segment_start = u64::from(segment.virtual_address) - analysis.identity.image_base;
-        let segment_end = segment_start + u64::from(segment.memory_size);
+        let segment_start = segment.virtual_address - analysis.identity.image_base;
+        let segment_end = segment_start + segment.memory_size;
         rva >= segment_start && end <= segment_end
     });
     let Some(segment) = segment else {
